Abstract CSV generalization logic
